requirements.txt
aniso8601==1.1.0
click==6.6
Flask==0.11.1
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
pytz==2016.6.1
RPi.GPIO==0.6.2
six==1.10.0
Werkzeug==0.11.11
